tests pass after SmartParse refactor
