+++ build.sbt
@@ -13,7 +13,7 @@ maxErrors := 10
 ThisBuild / scalaVersion  := scalaVer
 
 lazy val projectName = "uni"
-ThisBuild / version       := "0.6.2"
+ThisBuild / version       := "0.7.0"
 ThisBuild / versionScheme := Some("semver-spec")
 
 ThisBuild / organization         := "org.vastblue"
