+++ jsrc/minimizeQuotes.sc
@@ -1,6 +1,6 @@
 #!/usr/bin/env -S scala-cli shebang -Wunused:imports -Wunused:locals -deprecation
 
-//> using dep org.vastblue:uni_3:0.6.2
+//> using dep org.vastblue:uni_3:0.7.0
 
 import uni.* // {Path, Paths}
 import uni.io.* // {FastCsv, CsvWriter, Delimiter}
