+++ src/test/scala/uni/time/@
@@ -12,16 +12,6 @@ class ChronoParseTests extends FunSuite {
   // --------------------------------
   // parse testDates against expected
   // --------------------------------
-  test(s"ampm") {
-    val teststr = "04/12/1992 01:58 PM"
-    val pDate: LocalDateTime = parseDateChrono(teststr)
-    val pds2   = pDate.toString("yyyy-MM-dd'T'HH:mm:ss")
-    val expectedStamp = "1992-04-12T13:58:00"
-    assertEquals(pds2, expectedStamp,
-      s"Input: $teststr\nOutput: $pDate"
-    )
-  }
-
   for (((teststr, expectedTimestamp), index) <- testDatesExpected.zipWithIndex) {
     test(s"chronoParse should properly parse string [$teststr]") {
       val pDate: LocalDateTime = parseDateChrono(teststr)
